requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-search-documents>=11.5.3
azure-core>=1.32.0
azure-cosmos>=4.9.0
azure-ai-textanalytics>=5.3.0
azure-cognitiveservices-speech>=1.34.1
python-dotenv>=1.0.1
requests>=2.31.0
openai>=1.23.2
numpy
pandas
fastapi
uvicorn